Remove the TextTagAutoGenerateSection.
